[main] Extract embedded bootmgr.exe from within bootmgr, if provided
The bootmgr.exe binary is irritatingly difficult to obtain; it's not
present as a standalone file on either the Windows installation media
or on the hard disk of a typical Windows system, and must therefore be
extracted from another source (e.g. by using ImageX to extract it from
a copy of boot.wim).

The bootmgr binary, which contains a compressed copy of bootmgr.exe,
is easier to obtain.  Add support for extracting the required
bootmgr.exe image from within bootmgr, in order to simplify end-user
instructions.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>