[build] Ensure stacks are at edges of used memory
Under VMware, memory above %esp somehow gets overwritten, leading to
corrupt internal data structures and unpredictable behaviour.

Ensure that the protected-mode stack is at the top of our used memory,
and that the real-mode stack is at the bottom.  This keeps all of the
important data structures nicely contained within a region of memory
that is exclusively ours.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>