[main] Mark all VESA video modes as unsupported
There is a known bug in some common versions of bootmgr (including
those shipped with Windows 8 and Windows Server 2012) which prevents
it from falling back to text mode to display an error message if it
discovers that it does not have the fonts required to display the
error message in a graphical mode.

Work around this bug by intercepting all calls to INT 10,4f01 ("get
VBE mode information") and marking all video modes as "unsupported".
This seems to be sufficient to force bootmgr to remain in text mode
and display error messages correctly.

Note that the simpler approach of intercepting INT 10,4f00 and setting
CF to indicate the total absence of the VESA BIOS Extensions does not
work: bootmgr then remains in text mode but fails to display any
messages at all.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>