[efi] Block first attempt to open EFI_GRAPHICS_OUTPUT_PROTOCOL
There is a known bug in some common versions of bootmgfw.efi
(including those shipped with Windows 8 and Windows Server 2012) which
prevents it from falling back to text mode to display an error message
if it discovers that it does not have the fonts required to display
the error message in a graphical mode.

Work around this bug by failing the first call to OpenProtocol() for
EFI_GRAPHICS_OUTPUT_PROTOCOL.  This seems to be sufficient to force
bootmgfw.efi to remain in text mode and display error messages
correctly.

Note that we cannot fail all attempts to open the protocol, since that
will prevent the OS from booting.

This is the UEFI equivalent of commit 32ed9d1 ("[main] Mark all VESA
video modes as unsupported").

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>