[build] Place .stack section at 0x40000
bootmgr.exe hard-codes the address 0x30000 for use as a buffer
accessible by real-mode code.  Commit 70dbdc ("[libc] Increase stack
size to avoid hard-coded addresses in bootmgr.exe") increased the
stack size to 256kb in order to avoid trampling on this region.

Reduce the stack size back to 64kB and place the stack explicitly via
the linker script.  This allows us to verify that the remainder of the
binary does not cross the 0x30000 threshold.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>