[build] Avoid using objcopy to translate between architectures
The build process for %.i386.x86_64.o objects currently uses objcopy
to translate from a 32-bit object to a 64-bit object.  This seems not
to be reliable across a wide variety of binutils versions; in
particular, the use of .rel.* versus .rela.* relocation sections
causes issues.

Use an alternative approach, in which we first use gcc to generate
32-bit assembly, then prefix that with a .code32 directive and feed it
to the assembler to generate a 64-bit object.

This has the beneficial side-effect of allowing constants to be
merged, which slightly reduces the final binary size.

Reported-by: Octavio Schroeder <octavio.schroeder@gmail.com>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>