[build] Generate correct code for systems with more than 2GB of RAM
We currently build position-dependent code with the default code model
(equivalent to -mcmodel=small).  This model assumes that all symbols
lie in the lower 2GB of memory space.  This assumption is not correct
on a UEFI system, where we may be relocated to any address.

Fix by building position-independent code with hidden visibility.
This results in code which assumes only that the final binary will be
no larger than 2GB (which is correct).  Symbol references within code
sections will use 32-bit offsets relative to the instruction pointer,
and symbol references within data sections will be full 64-bit
pointers anyway.  This matches the approach taken by iPXE.

Reported-by: Brandon Penglase <bpenglase-ipxe@spaceservices.net>
Debugged-by: Brandon Penglase <bpenglase-ipxe@spaceservices.net>
Tested-by: Brandon Penglase <bpenglase-ipxe@spaceservices.net>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>