[efi] Force correct DeviceHandle on loaded bootmgfw.efi image
Some EFI/UEFI implementations seem to fail to fill in the DeviceHandle
for our loaded bootmgfw.efi image.  It is plausible that these
implementations fill in the DeviceHandle only if loading the image
from a device path (rather than directly from a memory buffer).

Work around this problem by filling in DeviceHandle if the firmware
leaves it empty, or uses the wrong device.

This mirrors iPXE commit 79419a1 ("[efi] Fill in loaded image's
DeviceHandle if firmware fails to do so").

Reported-by: Brandon Penglase <bpenglase-ipxe@spaceservices.net>
Tested-by: Brandon Penglase <bpenglase-ipxe@spaceservices.net>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>