[qa] Add a mechanism for running automatic wimboot tests
Create a test framework that will spin up a temporary virtual machine
configured to boot via iPXE and wimboot into a Windows PE environment,
and verify that the virtual machine correctly reaches the point of
running Windows applications.

Virtual machines are created via libvirt using QEMU's user-mode
networking, and can be run as an unprivileged user.

Boot files are served from a temporary local HTTP server running on an
unprivileged port and accessible only to localhost.

A QR code containing a random UUID is generated by the test framework
and added to wimboot's virtual filesystem along with a winpeshl.ini
that causes the QR code to be displayed once Windows has started up.
The test framework takes screenshots of the virtual machine once per
second and looks for the QR code to determine a successful test
result.

An interactive viewer is available when running locally.  The captured
wimboot output along with the final screenshot from each test is saved
to allow for debugging of test runs performed in a headless system
(e.g. as part of a CI workflow).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>