[ci] Add GitHub action for running QA tests
The QA tests require access to Windows boot images that unfortunately
cannot be downloaded from public sources, or be shared with other
GitHub users.

Fetch the Windows boot images via rsync using an SSH key stored as a
repository secret, and cache the fetched images within GitHub for
reuse.

Construct the cache key using a repository secret and a hash of the
expected file list.  The use of the repository secret prevents the
cache contents from being exposed to pull requests (and allows for the
cache to be effectively flushed manually by changing the secret).  The
use of the hash ensures that the cache is automatically refreshed
whenever the source file list changes.  The cache key is constructed
to allow for a cache restore key that can fetch an out-of-date version
of the cache, thereby allowing rsync to be used effectively to
minimise data transfer volumes.

The cache is refreshed daily to avoid excessive startup delays and
data transfer volumes.

Skip the entire workflow if the cache key secret is not defined, to
avoid noisy failures in forked repositories or on pull requests.

Limit tests to running on the "qa" branch only, at least until the
cost and reliability of these tests is better known.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>