[prefix] Allow initrd to be placed anywhere in 32-bit memory
As mentioned in commit c60402b ("[prefix] Limit initrd to the low 2GB
of memory"), some (32-bit) versions of Windows assume the ability to
use fixed addresses above 0x80000000.  We currently limit the initrd
to being placed below 2GB in order to avoid this potential issue.

This imposes an upper limit of ~2GB on the initrd size.  This limit
was previously irrelevant, since there was already a limit of ~2GB due
to the requirement to have enough space in 32-bit memory to hold two
copies of the .wim image (one accessed by wimboot, one for the Windows
ramdisk).

As of commit 1b2cb34 ("[paging] Remap initrd to physical memory above
4GB if possible") this latter limit no longer applies, at least in
systems with more than 4GB of RAM.

Allow the initrd to be placed anywhere in the 32-bit address space, in
order to allow for larger initrds on systems with sufficient RAM.

We first attempt to relocate below 2GB, failing gracefully if
insufficient RAM is available.  In the common case that the initrd is
only a few hundred MB, this will result in the initrd being placed
below 2GB and thereby avoiding problems with versions of Windows that
assume the ability to used fixed addresses above 0x80000000.

After parsing (and potentially expanding) the initrd, we then attempt
to relocate it above 4GB, failing gracefully if PAE is not supported
or if insufficient RAM is available.

The net result is that:

- For a system with PAE and sufficient RAM, the initrd will be placed
  above 4GB.  This avoids the problematic fixed addresses above
  0x80000000 and allows for an initrd larger than 2GB.

- For a system lacking PAE or with insufficient RAM, the initrd will
  be placed below 2GB if at all possible.  This avoids the problematic
  fixed addresses above 0x80000000.

- For a system lacking PAE or sufficient RAM and with a very large
  initrd, the boot will fail with a Windows error complaining about
  insufficient memory to create a ramdisk.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>