[wim] Always process .wim image file last
The order of input files is undefined on a UEFI system, which could
potentially lead to a situation in which a file automatically
extracted from a .wim image takes priority over a file provided
explicitly.

This issue has not arisen in practice since wimboot is almost always
loaded by iPXE, and iPXE happens to preserve the order of downloaded
files within the virtual filesystem.

Fix by deferring processing of the .wim file until after all other
input files have been read.

Modify the automated test script to place the .wim download in between
two files that are both required for the test to pass.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>