[main] Fix extraction of bootmgr.exe from Windows 10 bootmgr
The Windows 10 version of bootmgr includes an XCA-compressed embedded
copy of bootmgr.exe at an offset that is 8-byte aligned rather than
16-byte aligned.  Update the scan to search on 8-byte boundaries
instead of 16-byte boundaries.

Retain the 16-byte boundary requirement for LZNT1-compressed embedded
copies of bootmgr.exe, since fewer sanity checks are available for
LZNT1 and so the scan is more likely to hit false positives.

Update the test cases to include the ability to extract bootmgr.exe
from the Windows 10 version of bootmgr.

Reported-by: Steve Si <steve@easy2boot.com>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>