[cmdline] Fix debug printing of command line
Commit 9c576c7 ("[cmdline] Add "quiet" option to suppress debug
output") deferred the printing of the command line until after parsing
the "quiet" command-line argument, in order to allow the printing of
the command line to itself be inhibited.  Since the act of parsing the
command line will replace the argument separator characters with NULs,
this has the unwanted side effect of causing only the first
command-line parameter to be printed.

Fix by restoring the characters that were replaced by NULs during
parsing.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>