[build] Do not attempt to translate R_X86_64_32S relocation type
The sign-extending R_X86_64_32S relocation types do not map to a
four-byte PE relocation record, since the latter does not perform sign
extension.

We do not currently require these relocation types, since wimboot code
is not built with -mcmodel=kernel and so does not use addresses within
the negative 2GB of address space.

Remove the spurious and unused processing for this relocation type.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>