[build] Allow for variation in pesign output format
The output from "pesign -h" has varied in format over time.  The most
recent variation (in pesign 116 and later) swaps the order of the
filename and hash value, to provide consistency with the output order
from other checksumming tools such as sha256sum.

Use a regular expression to extract only the portion of the output
looking like a checksum value, thereby accommodating any version of
pesign released to date.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>