[efi] Use separate .text and .data PE sections
The .payload section currently contains all of the runtime executable
code, read-only data, and writable data.  It is therefore marked as
both writable and executable.  This is no longer permitted in binaries
signed for UEFI Secure Boot.

Fix by splitting the .payload section into separate .text and .data
sections.

The split is necessarily non-trivial, since the hybrid BIOS/UEFI
binary layout imposes the additional requirement that the 16-bit BIOS
executable code, initialised-data, and uninitialised-data sections are
all placed within the first 64kB of the RVA address space.  To
accommodate this, we currently place .bss16 first within the address
space, with .text16, .rodata16, and .data16 following immediately
afterwards at the start of the .payload section.

In the new split layout, we place the .text section after the .data
section, and place .text16, .rodata16, and .data16 at the start of the
.data section.  The .data section does therefore contain executable
code.  However, this is 16-bit BIOS code that can never be executed
under UEFI anyway, and may therefore be treated as opaque (and unused)
data from the perspective of a UEFI PE binary.

We do not further split .data into read-only and writable portions,
since there is no space to insert an extra PE section header before
the fixed-offset bzImage header.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>