[build] Align sections to page boundaries
Allow PE section attributes to be used to inform page-level protection
mechanisms such as W^X by aligning all sections to a page boundary.

Aligning the payload to a page boundary causes it to no longer
immediately follow the bzImage setup code.  Allow for this padding
when locating the start of the loaded-high payload to be copied to the
runtime address.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>