[build] Remove unnecessary padding after bzImage prefix
Section virtual addresses must be aligned to page boundaries in order
to support page-level protection mechanisms.  Section file offsets and
lengths do not need to be aligned to page boundaries.

Remove the padding of the bzImage prefix to reduce the size of the
overall binary.  Retain the padding within other sections, since the
hybrid BIOS/UEFI binary layout requires that the remainder of the file
(after the bzImage prefix) may be loaded into memory as a single
contiguous blob with no runtime relocation.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>