[build] Attempt to respect external tool choices
Attempt to respect external tool choices via variables such as $(CC),
$(AS), $(LD), $(AR), etc.

Cross-compiling versions of gcc are typically installed as
e.g. aarch64-linux-gnu-gcc rather than aarch64-linux-gnu-cc, and so a
naive expansion of $(CROSS_arm64)$(CC) would not produce the correct
binary name in the common case that $(CC) is set to "cc".  Work around
this quirk by redefining $(CC) as "gcc" if the detected compiler
claims to be gcc (i.e. defines the __GNUC__ macro).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>