[build] Support building with the clang compiler
Add support for building with the clang compiler as an alternative to
gcc.  Building with clang still requires the use of GNU binutils, due
to the assorted bits of assembler and linker trickery required to
produce a hybrid 32-bit/64-bit binary.

For x86_64, clang will access the __stack_chk_guard symbol via the
global offset table, using an R_X86_64_GOTPCRELX relocation.  The
linker relaxes this to an R_X86_64_32S relocation, which is not
supportable since the final PE executable may be relocated anywhere
within the 64-bit address space.  We cannot instruct the linker to
create a position-independent executable, since it will then complain
about impossible relocations in the 16-bit sections.  Instead, we
inhibit link-time relaxations so that the linked executable still
contains the R_X86_64_GOTPCRELX relocation, which elf2efi can then
ignore since it is PC-relative.

The address-significance table emitted by clang is not understood by
the GNU assembler, so add -fno-addrsig if applicable.

Add explicit register size modifiers for the ARM64 byte-swapping
inline assembly, since these seem to be required by clang.

Add -Wno-unused-command-line-argument to the global CFLAGS when using
clang, since otherwise many compiler invocations will fail (including
the compiler invocations used to determine the appropriate compiler
flags to be used).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>