[build] Fix building with newer binutils
Newer versions of the GNU assembler (observed with binutils 2.41) will
complain about the ".arch i386" in files assembled with "as --64",
with the message "Error: 64bit mode not supported on 'i386'".

Fix by moving ".arch i386" below ".code32", so that the assembler is
no longer expecting 64-bit instructions to be used by the time that
the ".arch i386" directive is encountered.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>