[qa] Enforce a virtual CPU capable of booting Windows 11
The default "qemu64" CPU model lacks support for some instructions
that are presumed to exist by newer versions of Windows 11, with the
result that the virtual machine will reboot almost immediately.

Fix by adding an explicit CPU declaration to the machine description
and specifying that the "popcnt" and "sse4.2" features are required.

The "qemu64" model requires support for the AMD-only "svm" feature,
which will not be available when using KVM on an Intel-based host.
When no explicit CPU declaration is present, libvirt seems to select a
"qemu64" CPU with the "svm" feature disabled if necessary.  When an
explicit declaration is present, this no longer happens automatically
and we must therefore explicitly mark "svm" as not required.  Choose
to mark it as explicitly disabled, to minimise the differences in
behaviour between running tests on Intel-based and AMD-based hosts.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>