[qa] Enable KVM for 64-bit tests in GitHub runners
As of some time in 2023, KVM may be supported in the public GitHub
runners, but the default permissions in Ubuntu do not allow the runner
user to access /dev/kvm.

Fix the udev rule for /dev/kvm to ensure that the runner user can
create KVM-accelerated virtual machines if supported by the (virtual)
hardware.

Use KVM only for the 64-bit tests, since experimentation shows that
the 32-bit tests will run incredibly slowly on some (but not all) of
the GitHub runners.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>