[qa] Switch to using "q35" machine type for Windows 11 UEFI test
The default "i440fx" machine type has no secure flash device and so
cannot be used for UEFI Secure Boot.  Allow the machine type to be
specified in the test definition, and switch to the "q35" machine type
for the Windows 11 UEFI test to allow for the possibility of enabling
Secure Boot.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>