[efi] Automatically select a suitably signed next stage bootloader
Windows 11 24H2 and newer include two versions of bootmgfw.efi: one
named bootmgfw.efi and signed with the existing "Microsoft Windows
Production PCA 2011", and the other named bootmgfw_EX.efi and signed
with the new "Windows UEFI CA 2023".

On a system that trusts both certificates, either bootloader will
work.  On an old system that does not recognise the newer "Windows
UEFI CA 2023" certificate, only bootmgfw.efi will be usable if Secure
Boot is enabled.  On a system that has had the older "Microsoft
Windows Production PCA 2011" certificate explicitly disabled, only
bootmgfw_EX.efi will be usable if Secure Boot is enabled.

Introspecting the system security policy to determine which
certificates or hashes are trusted (or blocked) is non-trivial and is
likely to be unreliable in practice, and so there is no viable way to
know in advance which bootloader ought to work.  We therefore attempt
bootmgfw_EX.efi first, falling back to bootmgfw.efi if the call to
LoadImage() fails.

To minimise surprise, we attempt extracting the two files from the WIM
image only if neither file was provided explicitly (e.g. via an iPXE
"initrd" command).  If any file is provided explicitly using the
default boot filename (e.g. "BOOTX64.EFI") then we will attempt
booting that file and only that file.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>